added actions to separate folders
